+++ data-raw/conditionWest.r
@@ -65,6 +65,9 @@ runnotes <- "This is the first conditioning on the west coast."
  datdir <- "c:/Users/User/DropBox/A_code/aMSEUse/conddata/sspmrun"
  resdir <- "c:/Users/User/DropBox/A_code/aMSEUse/conddata/sspm"
  dirExists(datdir,make=TRUE,verbose=TRUE)
+ source("C:/Users/User/Dropbox/A_Code/aMSE/data-raw/candidatefunctions.R")
+
+
  # You now need to ensure that there is a control.csv, zone1.csv
  # and region1.csv file in the data directory plus some data .csv files
  ctrl <- checkctrldat(datdir,ctrlfile="control.csv")
@@ -79,58 +82,59 @@ runnotes <- "This is the first conditioning on the west coast."
  product <- out$product     # important bits usually saved in resdir
  # did the larval dispersal level disturb the equilibrium?
  regDe <- testequil(zoneC,zoneD,glb)
-
- # do the depletions
  origzoneD <- zoneD
+ nSAU <- glb$nSAU
+ yrs <- zone1$histyr[,"year"]
+ histCE <- zone1$histCE
+ pickyr <- which(yrs %in% 1992:2019)
+ # do the depletions
+
  zoneD <- origzoneD
- origdepl <- c(0.6,0.6,0.7,0.6,0.6,0.6,0.6,0.6)
+ constants["initdepl",] <- c(0.85,0.75,0.8,0.9,0.725,0.95,0.75,0.9)
+ origdepl <- constants["initdepl",] #c(0.6,0.6,0.7,0.6,0.6,0.6,0.6,0.6)
  zoneDD <- depleteSAU(zoneC,zoneD,glb,depl=origdepl,product,len=15)
- zprop <- getzoneprops(zoneC,zoneDD,glb,year=1)
- round(zprop,4)
- zoneD <- zoneDD
+ if (ctrl$hcrname == "historicalC") zoneD <- dohistC(zoneDD,zoneC,zone1,glb)
+ predCE <- zoneD$cpue; rownames(predCE) <- yrs
+ pCE <- predCE
+ for (i in 1:nSAU) pCE[,i] <- predCE[,i]/mean(predCE[pickyr,i])
+ ssq <- rep(0.0,nSAU)
+ for (i in 1:nSAU) ssq[i] <- sum((log(histCE[,i]) - log(pCE[pickyr,i]))^2,na.rm=TRUE)
+ rbind(constants["initdepl",],round(ssq,5))
+ sum(ssq)
+ plotCPUE(pCE,histCE,pickyr,1:8)
+
+
+
+ plotSAUdepl(zoneD$deplsB)
+ zprop <- getzoneprops(zoneC,zoneD,glb,year=nyrs)
+ rbind(constants["initdepl",],round(zprop["SpBDepl",1:8],4))
+
 
  nSAU <- glb$nSAU
- str(zone1,max.level = 1)
- if (ctrl$hcrname == "historicalC") {
-   histC <- zone1$histCatch
-   histCE <- zone1$histCE
-   histyr <- zone1$histyr
-   yrs <- histyr[,"year"]
-   histLML <- histyr[,"histLML"]
-   nyrs <- length(yrs)
-   for (yr in 2:nyrs) {  # yr=2
-     year <- yrs[yr]
-     lml <- histLML[yr]
-     catchpop <- histC[yr,]
-     zoneD <- oneyearC(zoneC=zoneC,zoneD=zoneD,Ncl=glb$Nclass,
-                       catchp=catchpop,year=yr,sigmar=1e-08,
-                       npop=glb$nSAU,movem=glb$move)
-   }
+ yrs <- zone1$histyr[,"year"]
+ nyrs <- length(yrs)
+ predCE <- zoneD$cpue; rownames(predCE) <- yrs
+ histCE <- zone1$histCE
+ pickyr <- which(yrs %in% 1992:2019)
+ pCE <- predCE
+ for (i in 1:nSAU) pCE[,i] <- predCE[,i]/mean(predCE[pickyr,i])
+ ssq <- 0.0
+ for (i in 1:nSAU) ssq <- ssq + sum((histCE[,i] - pCE[pickyr,i])^2,na.rm=TRUE)
+ ssq
 
- }
- zpropF <- getzoneprops(zoneC,zoneD,glb,year=nyrs)
- round(zpropF,4)
 
- outB <- zoneD$deplsB
- plotprep(width=7,height=6,newdev = FALSE)
- parset(plots=c(3,1))
- ymax <- getmax(outB[,c(1,2,3,8)])
- plot(yrs,outB[,1],type="l",lwd=2,ylim=c(0,ymax),panel.first=grid(),ylab="6,7,8,13")
- for (i in c(2,3,8)) lines(yrs,outB[,i],lwd=2,col=i)
- legend("topright",legend=c(6,7,8,13),col=c(1,2,3,8),lwd=3,bty="n",cex=1.0)
- ymax <- getmax(outB[,c(4,5,7)])
- plot(yrs,outB[,4],type="l",lwd=2,ylim=c(0,ymax),col=4,panel.first=grid(),
-      ylab=c("9,10,12"))
- for (i in c(5,7)) lines(yrs,outB[,i],lwd=2,col=i)
- legend("topright",legend=c(9,10,12),col=c(4,5,7),lwd=3,bty="n",cex=1.0)
- ymax <- getmax(outB[,6])
- plot(yrs,outB[,6],type="l",lwd=2,ylim=c(0,ymax),col=6,panel.first=grid(),
-      ylab="11")
- legend("topright",legend=c(11),col=c(6),lwd=3,bty="n",cex=1.0)
+
+ plotCPUE(pCE,histCE,pickyr,1:8)
 
 
 
 
 
- # deplete SAUs separately--------------------------------------------------
+
+
+
+
+
+
+